requirements.txt
openai==0.28.1
langchain==0.0.354
tiktoken==0.5.2
python-dotenv==1.0.0
pytest==7.4.3